requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-dotenv==1.4.2
django-filter==2.2.0
djangorestframework==3.10.1
djangorestframework-stubs==0.4.2
future==0.17.1
idna==2.8
pytz==2019.1
requests==2.22.0
responses==0.10.6
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
